requirements.txt
flask
openai
scikit-learn
python-dotenv
pandas